+++ credits.php
@@ -1,0 +1,123 @@
+<!-- credits.php -->
+<html>
+<head>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Credits</title>
+    <script src="Auto_Display.js"></script>
+    <noscript><link href="default_stylesheet.css"></noscript>
+</head>
+<body>
+    <div class="FOIL-Font-Styles--sans-serif">
+        <div class="FOIL-Navigation-Bar">
+            <FOIL-Font-Size-24>Credits </FOIL-Font-Size-24>
+            <a href="about.php" class="FOIL-button-small color-is-green">Back</a>
+        </div>
+
+		<div class="FOIL-Layout">
+			<FOIL-Font-Size-22> Tabletop.js </FOIL-Font-Size-22>
+            <li> <b> <a href="https://github.com/jsoma/tabletop" class="FOIL-button-small color-is-green">https://github.com/jsoma/tabletop</a> </b> </li>
+            <i>
+            <pre>
+            Copyright (c) 2012-2018 Jonathan Soma
+            
+            Permission is hereby granted, free of charge, to any person
+            obtaining a copy of this software and associated documentation
+            files (the "Software"), to deal in the Software without
+            restriction, including without limitation the rights to use,
+            copy, modify, merge, publish, distribute, sublicense, and/or sell
+            copies of the Software, and to permit persons to whom the
+            Software is furnished to do so, subject to the following
+            conditions:
+            
+            The above copyright notice and this permission notice shall be
+            included in all copies or substantial portions of the Software.
+            
+            THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+            EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+            OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+            NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+            HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+            WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+            FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+            OTHER DEALINGS IN THE SOFTWARE.
+            
+            </i>
+            </pre>
+            <ul></ul>
+		</div>
+
+        <div class="FOIL-Layout">
+			<FOIL-Font-Size-22> jQuery </FOIL-Font-Size-22>
+            <li> <b> <a href="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.min.js" class="FOIL-button-small color-is-green">https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.min.js</a> </b> </li>
+            <li> <b> <a href="" class="FOIL-button-small color-is-green"></a> </b> </li>
+            <i>
+            <pre>
+            Copyright OpenJS Foundation and other contributors, https://openjsf.org/
+            
+            Permission is hereby granted, free of charge, to any person obtaining
+            a copy of this software and associated documentation files (the
+            "Software"), to deal in the Software without restriction, including
+            without limitation the rights to use, copy, modify, merge, publish,
+            distribute, sublicense, and/or sell copies of the Software, and to
+            permit persons to whom the Software is furnished to do so, subject to
+            the following conditions:
+            
+            The above copyright notice and this permission notice shall be
+            included in all copies or substantial portions of the Software.
+            
+            THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+            EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+            MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+            NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+            LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+            OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+            WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+            </i>
+            </pre>
+            <ul></ul>
+        </div>
+
+       <div class="FOIL-Layout">
+			<FOIL-Font-Size-22>RiveScript-JS </FOIL-Font-Size-22>
+            <li> <b> <a href="https://unpkg.com/rivescript@latest/dist/rivescript.min.js" class="FOIL-button-small color-is-green">https://unpkg.com/rivescript@latest/dist/rivescript.min.js</a> </b></li>
+            <li> <b> <a href="https://github.com/aichaos/rivescript-js" class="FOIL-button-small color-is-green">https://github.com/aichaos/rivescript-js</a> </b></li>
+            <i>
+            <pre>
+            The MIT License (MIT)
+            
+            Copyright (c) 2020 Noah Petherbridge
+            
+            Permission is hereby granted, free of charge, to any person obtaining a copy
+            of this software and associated documentation files (the "Software"), to deal
+            in the Software without restriction, including without limitation the rights
+            to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+            copies of the Software, and to permit persons to whom the Software is
+            furnished to do so, subject to the following conditions:
+            
+            The above copyright notice and this permission notice shall be included in all
+            copies or substantial portions of the Software.
+            
+            THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+            IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+            FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+            AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+            LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+            OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+            SOFTWARE.
+            </pre>
+            </i>
+            <ul></ul>
+
+
+       <div class="FOIL-Layout">
+			<FOIL-Font-Size-22>FOIL-UI-Framework-Library </FOIL-Font-Size-22>
+            <li> <b> <a href="https://github.com/senalbulumulle/FOIL-UI-Framework-Library" class="FOIL-button-small color-is-green">https://github.com/senalbulumulle/FOIL-UI-Framework-Library</a> </b>
+            <i>
+            <pre>
+            :-)
+            </pre>
+            </i>
+            <ul></ul>
+       </div>
+</body>
+</html>